Extract src/icons.ts — the per-project glyph store
Prerequisite for the sidebar slice: renderSidebar needs projGlyph and
renderMini needs iconFor, so by seam rule 1 the callee moves down first
rather than the renderer reaching up for it.

Moved whole (two line ranges, guarded slice): the two localStorage maps
(discovered `cc-icons` and hand-picked `cc-custom-icons`, deliberately
separate), the ICON_CACHE_VERSION migration, iconFor, probeIcon,
clearIcon, pickCustomIcon, resetCustomIcon and projGlyph. openProjectFolder
sat in the middle of the block and stayed behind — it opens a folder, it
has nothing to do with icons.

It sits below the sidebar rather than inside it because four surfaces read
the same store: the sidebar rows and the mini-rail paint it, the palette
shows it beside a session, and the colour popover offers set/clear/reset.

Two seam-rule-2 hooks, per-callee rather than one host object (PLAN prefers
setters below ~4): setIconRenderSidebar / setIconRenderMini. Naming them
after the callees keeps the four `renderSidebar(); renderMini();` call
sites byte-identical, the same shape diffview.ts uses for closeFootMenus.

`customIcons` is exported because the colour popover reads it directly to
decide whether "Restore repo logo" shows — same as main.ts reading
`colorOverrides` out of state.ts.

No tests: a render-adjacent module, per the Phase-1 render-module rule.
Verified mechanical — the relocated block diffed against
`git show HEAD:src/main.ts` differs only by seven `export` keywords.
Gates: tsc --noEmit clean, 352 vitest pass. No Rust touched.
